# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Gym Multi-Tenant API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier: { type: string, example: owner@actionplus.com }
                password: { type: string, format: password }
      responses:
        '200': { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT